+++ web-components/aofl-list-option/documentation/routes.config.js
@@ -1,0 +1,2 @@
+window.aofljsConfig = window.aofljsConfig || {};
+window.aofljsConfig.routesConfig = {};