+++ web-components/aofl-select-list/documentation/js/main-layout-element/template.js
@@ -1,0 +1,4 @@
+export const template = (context, html) => html`
+  <slot id="page-header" name="page-header"></slot>
+  <slot id="page-content" name="page-content"></slot>
+`;