+++ web-components/aofl-select-list/documentation/routes/home-element/examples/dynamic/index.js
@@ -1,0 +1,10 @@
+context.numbers = [];
+context.selected = '';
+
+context.addToList = (e) => {
+  context.numbers = [...context.numbers, context.numbers.length + 1];
+};
+
+context.updateSelected = (e) => {
+  context.selected = e.target.value;
+};