+++ web-components/aofl-select-list/documentation/routes/home-element/examples/select-list/index.js
@@ -1,0 +1,4 @@
+context.selected = '';
+context.updateSelected = (e) => {
+  context.selected = e.target.value;
+};