+++ web-components/aofl-select-list/documentation/routes/home-element/examples/toggle/index.js
@@ -1,0 +1,4 @@
+context.toggled = '';
+context.updateSelected = (e) => {
+  context.toggled = e.target.value === 'On' ? 'show' : '';
+};